Move detection of HTTP CONNECT proxy from DNS resolver to client channel.
